+++ foam/__init__.py
@@ -1,4 +1,5 @@
 from .core import Foam
+from .postprocessing import VTK
 
 
 __doc__ = Foam.__doc__