+++ foam/postprocessing/__init__.py
@@ -1,0 +1,4 @@
+__all__ = ['VTK']
+
+
+from .core import VTK